### 2024-04-09 e48f27c
Copula Vignette File Update
