+++ vignettes/ParametricCopulaFamilies.Rmd
@@ -5,7 +5,7 @@ params:
 output:
   bookdown::html_vignette2:
     toc: true
-author: "Alan, Elif, Martin"
+author: "Alan Kuchinsky, Elif F. Acar, Martin Lysy"
 date: 2024-03-26
 bibliography: copulaReferences.bib
 link-citations: true
@@ -30,9 +30,9 @@ library(LocalCop)
 ```
 
 
-This document will discuss the implementations of the copula families seen in this package. That is, we will discuss the CDF, conditional CDF, and PDF of each of the families implemented. For the conditional CDF, most of the copula implemented are symmetric in $u$ and $v$, so it is sufficient to provide one conditional CDF, and a simple swap of the variables will provide the other conditional CDF. Where this is not the case, both conditional CDF's will be provided. The document will offer both the original-scale and the logarithm-scale version of each formula where applicable. It should also be noted that the implementation utilizes the logarithm-scale version where possible to increase accuracy.
+This document will discuss the implementations of the copula families seen in this package. That is, we will discuss the CDF, conditional CDF, also known as the h-functions in uniform scale, and PDF of each of the families implemented. For the conditional CDF, most of the copula implemented are symmetric in $u$ and $v$, so it is sufficient to provide one conditional CDF, and a simple swap of the variables will provide the other conditional CDF. Where this is not the case, both conditional CDF's will be provided. The document will offer both the original-scale and the logarithm-scale version of each formula where applicable. It should also be noted that the implementation utilizes the logarithm-scale version where possible to increase accuracy.
 
-## Gaussian
+## Gaussian Copula
 
 The Gaussian copula family has a single parameter, which is the correlation between the two normal variables, $\rho \in (-1, 1)$. That is, the correlation matrix is:
 
@@ -58,7 +58,7 @@ Whose logarithm-scale is simplifiable and is presented below:
 $$\mathcal{c}(u, v; \rho) = -\frac{1}{2}\ln(1 - \rho^2) - \frac{1}{2} \frac{\rho^2  (\Phi^{-1}(u))^2 - 2 \rho \Phi^{-1}(u) \Phi^{-1}(v) + \rho^2 (\Phi^{-1}(v))^2 } {1-\rho^2}$$
 
 
-## Student
+## Student Copula
 
 The Student copula family has two parameters, the correlation between the two t-distributed variables, $\rho \in [-1, 1]$, and the degrees of freedom of both t-distributions, $\nu \in \mathbb{R}$, $\nu > 0$.
 
@@ -77,63 +77,8 @@ Similar to the CDF, the logarithm-scale is unsimplifiable, so it is omitted. The
 
[…]
