### 2024-04-15 ade3147
shortened joss manuscript
