+++ vignettes/joss/references.bib
@@ -12,6 +12,7 @@
     author = {{Wood}, S.N},
     edition = {2},
     publisher = {Chapman and Hall/CRC},
+    doi = {10.1201/9781315370279}
   }
 
 @Article{kristensen.etal16,
@@ -33,6 +34,7 @@
 	date-modified = {2024-03-04 04:18:39 -0600},
 	journal = {Econometrics and Statistics},
 	pages = {181--197},
+	doi = {10.1016/j.ecosta.2019.03.002},
 	title = {Dynamic vine copula models for multivariate time series data},
 	volume = {12},
 	year = {2019}}
@@ -43,7 +45,8 @@
 	date-modified = {2024-02-24 22:46:46 -0600},
 	journal = {Biometrics},
 	number = {2},
-	pages = {445-453},
+	pages = {445--453},
+	doi = {10.1111/j.1541-0420.2010.01472.x},
 	title = {Dependence calibration in conditional copulas: A nonparametric approach},
 	volume = {67},
 	year = {2011}}
@@ -53,11 +56,11 @@
 	date-added = {2024-02-24 22:40:49 -0600},
 	date-modified = {2024-02-24 22:46:52 -0600},
 	journal = {Electronic Journal of Statistics},
-	pages = {2822-2850},
+	pages = {2822--2850},
+	doi = {10.1214/13-EJS866},
 	title = {Statistical testing of covariate effects in conditional copula models},
 	volume = {7},
-	year = {2013},
-	bdsk-url-1 = {http://dx.doi.org/10.1111/j.1541-0420.2010.01472.x}}
+	year = {2013}}
 
 
 @Manual{localcop,
@@ -85,6 +88,7 @@
   number = {4},
   pages = {1--21},
   url = {https://www.jstatsoft.org/v21/i04/},
+  doi = {10.18637/jss.v021.i04}
 }
 
 @Article{copula3,
@@ -96,6 +100,7 @@
   number = {9},
   pages = {1--20},
   url = {https://www.jstatsoft.org/v34/i09/},
+  doi = {10.18637/jss.v034.i09}
 }
 
 @Article{copula4,
@@ -107,6 +112,7 @@
   number = {9},
   pages = {1--20},
   url = {https://www.jstatsoft.org/v39/i09/},
+  doi = {10.18637/jss.v039.i09}
 }
 
 
