### 2024-04-26 #6560 @EFAcar
We could suggest tnagler and AlexisDerumigny, the authors of alternative software for fitting conditional copulas.
