### 2024-05-06 #6560 @lrnv
👋 @RhysPeploe, @fartist, @FranjoIM, @mingzehuang, @ahoshiyar would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: [joss.readthedocs.io/en/latest/review_criteria.html](https://joss.readthedocs.io/en/latest/review_criteria.html)
