### 2024-05-06 #6560 @tvatter
Hi,

I'm sorry but I just completed a number of reviews and I am backed up by
other things.

Best,
Thibault

On Mon, May 6, 2024 at 9:45 AM Oskar Laverny ***@***.***>
wrote:

> 👋 @RhysPeploe <https://github.com/RhysPeploe>, @fartist
> <https://github.com/fartist>, @FranjoIM <https://github.com/FranjoIM>,
> @mingzehuang <https://github.com/mingzehuang>, @ahoshiyar
> <https://github.com/ahoshiyar> would any of you be willing to review this
> submission for JOSS? We carry out our checklist-driven reviews here in
> GitHub issues and follow these guidelines:
> joss.readthedocs.io/en/latest/review_criteria.html
>
