### 2024-05-13 #6744 @editorialbot
[REVIEW]: LocalCop: An R package for local likelihood inference for conditional copulas
**Submitting author:** @EFAcar (<a href="http://orcid.org/0000-0003-2908-7691">Elif Acar</a>)
**Repository:** https://github.com/mlysy/LocalCop
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.0.2
