### 2024-05-17 mlysy/LocalCop#7 @mingzehuang
Community Guideline
Hi, @EFAcar , is there a community guideline about contributions? Since the review checklist has the requirement as follow:
"Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the software 2) Report issues or problems with the software 3) Seek support"
[openjournals/joss-reviews#6744](https://github.com/openjournals/joss-reviews/issues/6744)
