### 2024-05-23 mlysy/LocalCop#7 @EFAcar
Hi @mingzehuang, the `DESCRIPTION` file has a `BugReports` section linking to the issue tracker.  In addition, we added a Contributions section to the README. Please let us know if you'd like us to do add anything else for this.
