### 2024-05-27 mlysy/LocalCop#10 @mingzehuang
Documentation Style
Hi, @EFAcar. I checked the vignette of your package. It seems you didn't organize it according to the several parts required in the 'Documentation' section of the checklist. Do you have a package website, or do you mind rearranging your vignette according to the structure required by the checklist? [openjournals/joss-reviews#6744](https://github.com/openjournals/joss-reviews/issues/6744#)
