### 2024-05-27 mlysy/LocalCop#10 @EFAcar
Hi @mingzehuang.  Here is a list of the items in the Documentation checklist and where they can be found in the package:

- *Statement of need:* Missing (currently only in the manuscript).
- *Installation instructions:* In the [README](https://github.com/mlysy/LocalCop/blob/joss/README.md) file.
- *Example usage:* In the [vignette](https://github.com/mlysy/LocalCop/blob/joss/vignettes/LocalCop-vignette.Rmd).
- *Functionality documentation:* All functions are fully documented in the [man](https://github.com/mlysy/LocalCop/tree/joss/man) folder, which can easily be browsed once the package is installed.
- *Automated tests:* In the [tests/testthat](https://github.com/mlysy/LocalCop/tree/joss/tests/testthat) folder.  These can be run after installing the package using `testthat::test_package("LocalCop")`.  **Note:** we've updated the `devtools::install_github()` command to install the package development version in the README to install the tests as well, so that the `test_package()` command above works.  The README now also contains a section explaining how to run the tests.
- *Community guidelines:* In the [README](https://github.com/mlysy/LocalCop/blob/joss/README.md).

So, would it be OK just to add a "statement of need" to the vignette to satisfy all the Documentation requirements?
