### 2024-05-29 mlysy/LocalCop#10 @mingzehuang
Sounds great to me @EFAcar !
Once you add the statement of need to the vignette, I'll close this issue:)
