### 2024-05-31 mlysy/LocalCop#10 @EFAcar
Hi @mingzehuang. The vignette now contains a "statement of need" in the second paragraph of the overview section.
