### 2024-08-05 mlysy/LocalCop#15 @AlexisDerumigny
Documentation mismatch for `ConvertPar`
Hi @EFAcar, I am checking the code in the `ConvertPar.R` file and it seems that the range of families mentioned in the documentation does not match the one actually implemented. The documentation says:

https://github.com/mlysy/LocalCop/blob/c9f16c287b87e5a4ea0bc8ddf93386cdb631d06e/R/ConvertPar.R#L19-L26

but it seems to me that a lot more families are actually implemented in the same file below (maybe all of them, in fact).
It could be nice to mention this in the documentation too.

https://github.com/openjournals/joss-reviews/issues/6744
