### 2024-08-05 mlysy/LocalCop#16 @AlexisDerumigny
Add link to `TMB` package in the `README`
Hi @EFAcar, in the `README`, I think it would be nice to add a hyperlink to the `TMB` package so that the reader can see check quickly what this is. (I did not know it and it looks very useful!)
Maybe in this sentence it could be good to add something like "using the TMB package" to be clearer for the reader.

https://github.com/mlysy/LocalCop/blob/c9f16c287b87e5a4ea0bc8ddf93386cdb631d06e/README.md?plain=1#L11

https://github.com/openjournals/joss-reviews/issues/6744
