### 2024-08-05 mlysy/LocalCop#15 @AlexisDerumigny
After reading more, I have the impression that this is because the rest of the code is not compatible with families outside 1-5.

Would you think it would be a lot of work to implement also the other families? In particular, I have the impression that adding the rotated ones would be quite similar to the existing code.

(corresponding to the idea "JOSS requires that software should be feature-complete", see https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort )
