+++ src/TMB/pgumbel.hpp
@@ -1,4 +1,4 @@
-#include "LocalCop/gumbel.hpp"
+#include "../../inst/include/LocalCop/gumbel.hpp"
 
 #undef TMB_OBJECTIVE_PTR
 #define TMB_OBJECTIVE_PTR obj
@@ -8,7 +8,7 @@ template <class Type>
     // R inputs
     DATA_VECTOR(u1);
     DATA_VECTOR(u2);
-    DATA_VECTOR(weights)
+    DATA_VECTOR(weights);
     PARAMETER_VECTOR(theta);
     // output
     vector<Type> lcdf = LocalCop::pgumbel(u1, u2, theta, 1);
