+++ src/TMB/pstudent.hpp
@@ -1,19 +1,20 @@
-#include "../../inst/include/LocalCop/gaussian.hpp"
+#include "../../inst/include/LocalCop/student.hpp"
 
 #undef TMB_OBJECTIVE_PTR
 #define TMB_OBJECTIVE_PTR obj
 
 template <class Type>
-Type dgaussian(objective_function<Type> *obj) {
+Type pstudent(objective_function<Type> *obj) {
   // R inputs
   DATA_VECTOR(u1);
   DATA_VECTOR(u2);
   DATA_VECTOR(weights);
   PARAMETER_VECTOR(theta);
+  PARAMETER_VECTOR(nu);
   // output
-  vector<Type> lpdf = LocalCop::dgaussian(u1, u2, theta, 1);
-  lpdf.array() *= weights.array();
-  return -lpdf.sum();
+  vector<Type> lcdf = LocalCop::pstudent(u1, u2, theta, nu, 1);
+  lcdf.array() *= weights.array();
+  return -lcdf.sum();
 }
 
 #undef TMB_OBJECTIVE_PTR
