+++ src/TMB/hstudent.hpp
@@ -1,4 +1,4 @@
-#include "LocalCop/student.hpp"
+#include "../../inst/include/LocalCop/student.hpp"
 
 #undef TMB_OBJECTIVE_PTR
 #define TMB_OBJECTIVE_PTR obj
@@ -8,7 +8,7 @@ Type hstudent(objective_function<Type> *obj) {
   // R inputs
   DATA_VECTOR(u1);
   DATA_VECTOR(u2);
-  DATA_VECTOR(weights)
+  DATA_VECTOR(weights);
   PARAMETER_VECTOR(theta);
   PARAMETER_VECTOR(nu);
   // output
