### 2024-08-22 f5b00a9
Update hstudent.hpp
