+++ src/TMB/pgaussian.hpp
@@ -1,4 +1,4 @@
-#include "../../inst/include/LocalCop/gaussian.hpp"
+#include "LocalCop/gaussian.hpp"
 
 #undef TMB_OBJECTIVE_PTR
 #define TMB_OBJECTIVE_PTR obj
