+++ R/CondiCopLikCV.R
@@ -35,8 +35,8 @@ CondiCopLikCV <- function(u1, u2, family, x, xind = 100,
     xind <- unique(round(seq(1, length(x), len = xind)))
   }
   # initialize eta and nu
-  if(!degree %in% 0:1) stop("degree must be 0 or 1.")
-  ## degree <- match.arg(degree)
+  .check_family(family)
+  .check_degree(degree)
   etaNu <- .get_etaNu(u1 = u1, u2 = u2, family = family,
                       degree = degree, eta = eta, nu = nu)
   ieta <- etaNu$eta
