+++ src/TMB/pclayton.hpp
@@ -4,17 +4,17 @@
 #define TMB_OBJECTIVE_PTR obj
 
 template <class Type>
-  Type pclayton(objective_function<Type> *obj) {
-    // R inputs
-    DATA_VECTOR(u1);
-    DATA_VECTOR(u2);
-    DATA_VECTOR(weights)
-    PARAMETER_VECTOR(theta);
-    // output
-    vector<Type> lcdf = LocalCop::pclayton(u1, u2, theta, 1);
-    lcdf.array() *= weights.array();
-    return -lcdf.sum();
-  }
+Type pclayton(objective_function<Type> *obj) {
+  // R inputs
+  DATA_VECTOR(u1);
+  DATA_VECTOR(u2);
+  DATA_VECTOR(weights);
+  PARAMETER_VECTOR(theta);
+  // output
+  vector<Type> lcdf = LocalCop::pclayton(u1, u2, theta, 1);
+  lcdf.array() *= weights.array();
+  return -lcdf.sum();
+}
 
 #undef TMB_OBJECTIVE_PTR
 #define TMB_OBJECTIVE_PTR this
