### 2024-09-03 mlysy/LocalCop#16 @mlysy
Thanks @AlexisDerumigny, we've implemented this suggestion in #18.
