### 2024-09-03 mlysy/LocalCop#15 @mlysy
Thanks @AlexisDerumigny, we've added the rotated versions of applicable copulas within families 1-5 in #18.
