+++ DESCRIPTION
@@ -3,10 +3,10 @@ Title: Local Likelihood Inference for Conditional Copula Models
 Version: 0.0.2
 Date: 2024-08-30
 Authors@R: 
-    c(person(given = "Elif",
+    c(person(given = "Elif Fidan",
              family = "Acar",
              role = "aut",
-             email = "elif.acar@umanitoba.ca"),
+             email = "eacar@uoguelph.ca"),
       person(given = "Martin",
              family = "Lysy",
 	     role = c("aut", "cre"),
