### 2024-09-05 fd6a2af
initial pkgdown attempt
- added "_" to joss vignette to prevent it from compiling
- removed "-vignette" from main vignette to be recognized by "Getting started"
- fails because we're using bookdown for numbering which doesn't seem to be supported.
