### 2024-09-06 6d31a8e
got pkgdown bookdown working with bootstrap 3
