+++ pkgdown.yml
@@ -1,0 +1,9 @@
+pandoc: 3.1.11
+pkgdown: 2.1.0
+pkgdown_sha: ~
+articles:
+  LocalCop: LocalCop.html
+last_built: 2024-09-06T01:20Z
+urls:
+  reference: https://mlysy.github.io/LocalCop/reference
+  article: https://mlysy.github.io/LocalCop/articles
