### 2024-09-10 e6ea66d
added CRAN package DOIs
