+++ vignettes/joss/references.bib
@@ -69,6 +69,7 @@
   year = {2024},
   note = {R package version 0.0.1},
   url = {https://CRAN.R-project.org/package=LocalCop},
+  doi = {10.32614/CRAN.package.LocalCop}
 }
 
 @Manual{copula1,
@@ -77,6 +78,7 @@
   year = {2023},
   note = {R package version 1.1-2},
   url = {https://CRAN.R-project.org/package=copula},
+  doi = {10.32614/CRAN.package.copula}
 }
 
 @Article{copula2,
@@ -123,6 +125,7 @@
   year = {2023},
   note = {R package version 2.5.0},
   url = {https://CRAN.R-project.org/package=VineCopula},
+  doi = {10.32614/CRAN.package.VineCopula}
 }
 
 
@@ -134,6 +137,7 @@ Structures and Vine Copulas},
   year = {2020},
   note = {R package version 0.0-7},
   url = {https://CRAN.R-project.org/package=gamCopula},
+  doi = {10.32614/CRAN.package.gamCopula}
 }
 
 
@@ -143,4 +147,5 @@ Structures and Vine Copulas},
   year = {2023},
   note = {R package version 0.1.3},
   url = {https://CRAN.R-project.org/package=CondCopulas},
+  doi = {10.32614/CRAN.package.CondCopulas}
 }
