### 2024-09-10 #6744 @EFAcar
Thank you @lrnv.  We've added DOIs for all CRAN packages (including the skipped ones) above.
