### 2024-09-12 be4286d
added source code to accepted version of joss paper
