+++ vignettes/joss/paper.md
@@ -69,22 +69,7 @@ link-citations: true
 
 -->
 
-```{r setup, include = FALSE}
-knitr::read_chunk("calculations.R")
-library(LocalCop)
-library(VineCopula)  # for simulating copula data
-library(readr)       # for reading in the copula table
-options(
-  kableExtra.latex.load_packages = FALSE
-)                    # for styling the copula table
-library(kableExtra)  # add packages manually for joss submission
-library(tidyr)       # for data manipulations
-library(dplyr)       #
-library(ggplot2)     # for plots
-library(ggpubr)      # 
-library(gamCopula)   # other packages for fitting
-library(CondCopulas) # conditional copula models
-```
+
 
 <!-- latex macros -->
 \newcommand{\R}{{\textsf{R}}}
@@ -143,91 +128,133 @@ where $(u_i, v_i, x_i)$ is the data for observation $i$, $\boldsymbol{x}_i = (1,
 
 In the following example, we illustrate the model selection/tuning and fitting steps for data generated from a Clayton copula with conditional Kendall $\tau$ displayed in \autoref{fig:copcomp}.  The CV metric for each combination of family and bandwidth are displayed in \autoref{fig:select1-plot}.
 
-```{r libs}
+
+``` r
+library(LocalCop)   # local likelihood estimation
+library(VineCopula) # simulate copula data
 ```
 
-```{r dgm1-setup}
-```
-```{r dgm1}
-<<dgm>>
-```
-```{r select1-precalc}
-<<select-precalc>>
-```
-```{r select1-calc, eval = params$do_calc, warning = FALSE}
-<<select-calc>>
-```
-```{r select1-save, include = FALSE, eval = params$do_calc}
-saveRDS(cvselect, file = "cvselect1.rds")
-```
-```{r select1-load, include = FALSE, eval = !params$do_calc}
-cvselect <- readRDS("cvselect1.rds")
-```
-```{r select1-plot, echo = FALSE, fig.height = 2, fig.width = 9, out.width = "100%", fig.cap = "Cross-validation metric for each combination of family and bandwidth."}
-<<select-plot>>
+
+``` r
+set.seed(2024)
+
+# simulation setting
+family <- 3                    # Clayton Copula
+n_obs <- 300                   # number of observations
+eta_fun <- function(x) {       # calibration function
+  sin(5*pi*x) + cos(8*pi*x^2)
+}
 ```
 
-```{r locfit1}
-<<locfit>>
-```
-```{r gamfit1, include = FALSE}
-<<gamfit>>
-```
-```{r condfit1, include = FALSE}
-<<condfit>>
-```
-```{r sumfit1, include = FALSE}
-sumfit1 <- tibble(
-  x = x0,
-  True = BiCopEta2Tau(family = family, eta = eta_fun(x0)),
-  LocalCop = tau_loc,
-  gamCopula = tau_gam,
-  CondCopulas = tau_cond
-)
-```
-```{r dgm2-setup, include = FALSE}
-```
-```{r dgm2}
-<<dgm>>
-```
-```{r select2-precalc}
-<<select-precalc>>
-```
-```{r select2-calc, eval = params$do_calc, warning = FALSE}
-<<select-calc>>
-```
-```{r select2-save, include = FALSE, eval = params$do_calc}
-saveRDS(cvselect, file = "cvselect2.rds")
-```
-```{r select2-load, include = FALSE, eval = !params$do_calc}
-cvselect <- readRDS("cvselect2.rds")
-```
-```{r locfit2}
-<<locfit>>
-```
-```{r gamfit2, include = FALSE}
-<<gamfit>>
-```
[…]
