+++ vignettes/joss/references.bib
@@ -9,15 +9,15 @@
 @Book{wood17,
     title = {Generalized Additive Models: An Introduction with {R}},
     year = {2017},
-    author = {S.N Wood},
+    author = {{Wood}, S.N},
     edition = {2},
     publisher = {Chapman and Hall/CRC},
+    doi = {10.1201/9781315370279}
   }
-
 
 @Article{kristensen.etal16,
   title = {{TMB}: Automatic Differentiation and {L}aplace Approximation},
-  author = {Kasper Kristensen and Anders Nielsen and Casper W. Berg and Hans Skaug and Bradley M. Bell},
+  author = {{Kristensen}, Kasper and {Nielsen}, Anders and {Berg}, Casper W. and {Skaug}, Hans and {Bell}, Bradley M.},
   journal = {Journal of Statistical Software},
   year = {2016},
   volume = {70},
@@ -29,78 +29,92 @@
 
 
 @article{ACL2019,
-	author = {Acar, Elif F. and Czado, Claudia and Lysy, Martin},
+	author = {{Acar}, Elif F. and {Czado}, Claudia and {Lysy}, Martin},
 	date-added = {2024-03-04 04:18:24 -0600},
 	date-modified = {2024-03-04 04:18:39 -0600},
 	journal = {Econometrics and Statistics},
 	pages = {181--197},
+	doi = {10.1016/j.ecosta.2019.03.002},
 	title = {Dynamic vine copula models for multivariate time series data},
 	volume = {12},
 	year = {2019}}
 
 @article{ACY2011,
-	author = {Acar, Elif F. and Craiu, Radu V. and Yao, Fang},
+	author = {{Acar}, Elif F. and {Craiu}, Radu V. and {Yao}, Fang},
 	date-added = {2024-02-24 22:40:49 -0600},
 	date-modified = {2024-02-24 22:46:46 -0600},
 	journal = {Biometrics},
 	number = {2},
-	pages = {445-453},
+	pages = {445--453},
+	doi = {10.1111/j.1541-0420.2010.01472.x},
 	title = {Dependence calibration in conditional copulas: A nonparametric approach},
 	volume = {67},
 	year = {2011}}
 
 @article{ACY2013,
-	author = {Acar, Elif F. and Craiu, Radu V. and Yao, Fang},
+	author = {{Acar}, Elif F. and {Craiu}, Radu V. and {Yao}, Fang},
 	date-added = {2024-02-24 22:40:49 -0600},
 	date-modified = {2024-02-24 22:46:52 -0600},
 	journal = {Electronic Journal of Statistics},
-	pages = {2822-2850},
+	pages = {2822--2850},
+	doi = {10.1214/13-EJS866},
 	title = {Statistical testing of covariate effects in conditional copula models},
 	volume = {7},
-	year = {2013},
-	bdsk-url-1 = {http://dx.doi.org/10.1111/j.1541-0420.2010.01472.x}}
+	year = {2013}}
 
 
+@Manual{localcop,
+  title = {LocalCop: LocalCop: Local Likelihood Inference for Conditional Copula Models},
+  author = {{Acar}, Elif F. and {Lysy}, Martin},
+  year = {2024},
+  note = {R package version 0.0.1},
+  url = {https://CRAN.R-project.org/package=LocalCop},
+  doi = {10.32614/CRAN.package.LocalCop}
+}
 
 @Manual{copula1,
   title = {copula: Multivariate Dependence with Copulas},
-  author = {Marius Hofert and Ivan Kojadinovic and Martin Maechler and Jun Yan},
+  author = {{Hofert}, Marius and {Kojadinovic}, Ivan and {M\"achler}, Martin and {Yan}, Jun},
   year = {2023},
[…]
