### 2024-09-20 #6744 @EFAcar
@lrnv, @mingzehuang, @AlexisDerumigny thank you for the helpful review comments which have much improved our submission.  Please find the post-review check list below:

    Thanks for asking us to double-check these.  These have been updated on the paper branch `mlysy/joss` after the post-review checks began.  The most recent manuscript file is [`vignettes/joss/paper.md`](https://github.com/mlysy/LocalCop/blob/joss/vignettes/joss/paper.md) on that branch.
    The post-review software has been released on [CRAN](https://cran.r-project.org/package=LocalCop).  The version number for the release is 0.0.2.
    The archive is posted on Zenodo with DOI https://doi.org/10.5281/zenodo.13801203.
