### 2024-09-20 #6744 @EFAcar
> @EFAcar Everytthing looks good, except that the Zenodo archive's title and the paper title must be the same, could you change one of them so that they match ?

@lrnv the title for the Zenodo archive is now corrected.
