+++ pkgdown.yml
@@ -1,9 +1,9 @@
 pandoc: 3.1.11
-pkgdown: 2.1.0
+pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
   LocalCop: LocalCop.html
-last_built: 2024-09-12T15:17Z
+last_built: 2024-09-23T01:32Z
 urls:
   reference: https://mlysy.github.io/LocalCop/reference
   article: https://mlysy.github.io/LocalCop/articles
