### 2024-09-23 #6744 @EFAcar
@lrnv based on [this review](https://github.com/openjournals/joss-reviews/issues/5358) the error code seems to be due to using displaymode with \dfrac. We now changed these to \frac. Hope this will fix the issue.
