### 2024-09-24 #6744 @crvernon
@EFAcar - this is looking good.  I just need you to address the following minor items before I accept for publication:

- [ ] LINE 85:  "LocalCop:" is repeated in the reference
- [ ] LINE 91:  The "a" in "archimedean" should be capitalized.  You can maintain capitalization in your references by using curly brackets around the characters you wish to maintain formatting for.

No need to conduct another release.

Let me know once these have been corrected and I will accept this for publication.
