### 2024-09-24 2d815a1
fixed typos in references
