+++ vignettes/joss/references.bib
@@ -64,7 +64,7 @@
 
 
 @Manual{localcop,
-  title = {LocalCop: LocalCop: Local Likelihood Inference for Conditional Copula Models},
+  title = {LocalCop: Local Likelihood Inference for Conditional Copula Models},
   author = {{Acar}, Elif F. and {Lysy}, Martin},
   year = {2024},
   note = {R package version 0.0.1},
@@ -106,7 +106,7 @@
 }
 
 @Article{copula4,
-  title = {Nested Archimedean Copulas Meet {R}: The {nacopula} Package},
+  title = {Nested {Archimedean} Copulas Meet {R}: The {nacopula} Package},
   author = {{Hofert}, Marius and {M\"achler}, Martin},
   journal = {Journal of Statistical Software},
   year = {2011},
