+++ pkgdown.yml
@@ -3,7 +3,7 @@ pkgdown: 2.1.1
 pkgdown_sha: ~
 articles:
   LocalCop: LocalCop.html
-last_built: 2024-09-23T01:32Z
+last_built: 2024-09-24T19:49Z
 urls:
   reference: https://mlysy.github.io/LocalCop/reference
   article: https://mlysy.github.io/LocalCop/articles
