### 2024-09-24 #6744 @EFAcar
@crvernon Thanks very much for catching these. They have now been fixed in this [PR](https://github.com/mlysy/LocalCop/pull/20).
